Refactoring: Extract metrics collector from StatelessCommitService
