+++ x-pack/plugin/stateless/src/test/java/org/elasticsearch/xpack/stateless/cache/SharedBlobCacheWarmingServiceTests.java
@@ -74,6 +74,7 @@ import org.elasticsearch.xpack.stateless.commits.StatelessCommitService;
 import org.elasticsearch.xpack.stateless.commits.StatelessCompoundCommit;
 import org.elasticsearch.xpack.stateless.commits.VirtualBatchedCompoundCommit;
 import org.elasticsearch.xpack.stateless.commits.VirtualBatchedCompoundCommitTestUtils;
+import org.elasticsearch.xpack.stateless.commits.metering.BccUploadMetricsCollector;
 import org.elasticsearch.xpack.stateless.engine.PrimaryTermAndGeneration;
 import org.elasticsearch.xpack.stateless.lucene.BlobCacheIndexInput;
 import org.elasticsearch.xpack.stateless.lucene.BlobStoreCacheDirectoryMetrics;
@@ -2216,8 +2217,8 @@ public class SharedBlobCacheWarmingServiceTests extends ESTestCase {
                         .filter(
                             measurement -> Double.compare(measurement.getDouble(), blobSpec.ratio()) == 0
                                 && measurement.attributes()
-                                    .get(StatelessCommitService.BCC_SIZE_ATTRIBUTE_KEY)
-                                    .equals(StatelessCommitService.bccSizeBucket(blobSpec.blobSize()))
+                                    .get(BccUploadMetricsCollector.BCC_SIZE_ATTRIBUTE_KEY)
+                                    .equals(BccUploadMetricsCollector.bccSizeBucket(blobSpec.blobSize()))
                         )
                         .count(),
                     is(1L)
@@ -2235,12 +2236,12 @@ public class SharedBlobCacheWarmingServiceTests extends ESTestCase {
                 .thenComparing(Map.Entry::getValue);
             assertThat(
                 requestedBytesMeasurements.stream()
-                    .map(m -> Map.entry(m.getLong(), m.attributes().get(StatelessCommitService.BCC_SIZE_ATTRIBUTE_KEY).toString()))
+                    .map(m -> Map.entry(m.getLong(), m.attributes().get(BccUploadMetricsCollector.BCC_SIZE_ATTRIBUTE_KEY).toString()))
                     .sorted(byValueThenBucket)
                     .toList(),
                 equalTo(
                     blobSpecs.stream()
-                        .map(spec -> Map.entry(spec.endOffset(), StatelessCommitService.bccSizeBucket(spec.blobSize())))
+                        .map(spec -> Map.entry(spec.endOffset(), BccUploadMetricsCollector.bccSizeBucket(spec.blobSize())))
                         .sorted(byValueThenBucket)
                         .toList()
                 )
